+++ src/edu/upenn/cit594/processor/AvgMktValue.java
@@ -2,12 +2,11 @@ package edu.upenn.cit594.processor;
 
 
 import edu.upenn.cit594.util.PropertyData;
-
 import java.util.List;
 
 public class AvgMktValue implements PropertyFunction {
     @Override
-    public double compute(List<PropertyData> properties) {
+    public int compute(List<PropertyData> properties) {
         double sum = 0;
         int count = 0;
         for (PropertyData p : properties) {
@@ -17,7 +16,7 @@ public class AvgMktValue implements PropertyFunction {
                 count++;
             }
         }
-        return count == 0 ? 0 : sum / count;
+        return count == 0 ? 0 : (int)(sum / count);
     }
 }
