+++ src/edu/upenn/cit594/processor/PropertyFunction.java
@@ -5,5 +5,5 @@ import java.util.List;
 import edu.upenn.cit594.util.PropertyData;
 
 public interface PropertyFunction {
-    double compute(List<PropertyData> properties);
+    int compute(List<PropertyData> properties);
 }
